journal: altered head/tail invariants.
The new invariant no longer requires that head/tail point at good
blocks, since that might not be possible to enforce. Instead, we only
require that they point at valid user-page indices.

We check before writing whether we have a place to put the head pointer,
and consumers are required to ask for a valid tail pointer before
reading from it.