journal: safe tail reuse.
We keep track of the tail at the time of the last sync event, and only
reuse the tail space when we're sure it couldn't lead to data loss in
the event of a power failure.

We assume that operations are fed by the user in an order which
guarantees consistency if we were to roll back to any point in the
sequence.