Fix for a bug leading to data loss under some circumstances.
If power was lost at a point where the journal head pointer was sitting
in the last checkpoint group of the NAND chip, then a resume would
correct wrap around the head pointer for the next write, but incorrectly
fail to increment the epoch counter.

This meant that any subsequent writes would have an epoch counter which
was lower than it should have been. If the journal didn't wrap around
again before the next resume, then the resume code would incorrectly
deduce that the pages at the start of the chip were older than those at
the end, and resume to an older checkpoint.

The resume code is modified to increment the epoch counter if a
wraparound occurs during resume. A similar bug may have occured in
skip_block() if the very last block in a chip was bad, and a similar fix
has been made.

Thanks to Sylvain Basset for reporting and diagnosing the bug.