Fix possible data loss after power failure when epoch is 0xff.
If a power failure occured prior to resume, and the current epoch is
0xff, and the NAND driver uses a checksum which is valid for
unprogrammed pages, then find_root() may erroneously pick a blank page
instead of a valid checkpoint page, because it looks only for a matching
epoch byte. We need to look for the magic word as well to avoid this.

Issue pointed out by gtsee on github.com.