journal: fix automatic root reset.
The automatic setting of the root to DHARA_PAGE_NONE on journal empty
was not occuring if the journal was only non-empty due to the existence
of bad blocks at the head. Since it's not practical to detect this, the
behaviour is altered to clear the root whenever it drops out of range.

Fortunately this situation is preempted by the map layer and should not
be causing a problem in practice anyway.

Issue reported by user dkoreshkov on Github:

    https://github.com/dlbeer/dhara/issues/19