journal: avoid querying past end during recovery.
cp_free() was misused here, as it was being called on individual
userpages instead of a checkpoint group as a whole. As a consequence, it
was reading past the end of the chip when invoked on a userpage within
ppc of the last page.

This was also O(n^2) in the size of checkpoint groups, although it
perhaps didn't matter in practice due to the small n and the fact that
it only ran once on startup.